broke benchmark and sanity checks into two files for history matching
